# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Go Chat MSA API
  version: "1.0.0"
  license:
    name: 무단 사용 금지
    url: https://choosealicense.com/no-permission/
  description: |
    Go Chat MSA 프로젝트의 REST API 명세서입니다.
    - REST API 에러는 RFC 9457 (Problem Details) 형식을 따릅니다. WebSocket 업그레이드 검증 오류는 일반 HTTP 오류 본문일 수 있습니다.
    - 보안을 위해 JWT(Access Token)와 HttpOnly Cookie(Refresh Token)를 혼합하여 사용합니다.
    - 웹소켓 연결은 1회용 단기 티켓 기반으로 수행됩니다.
servers:
  - url: /api
    description: Kubernetes Ingress를 통한 API Gateway 경로
...
paths:
...
  /auth/token/refresh:
    post:
      tags:
        - Auth
      summary: Access Token 재발급
      description: 보안 쿠키에 담긴 Refresh Token을 확인하여 새로운 Access Token을 발급합니다.
      operationId: refreshToken
      security:
        - refreshCookie: []
      responses:
        '200':
          description: 재발급 성공
          headers:
            Set-Cookie:
              description: 새로 갱신된 Refresh Token 쿠키 (RTR 적용)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "API 요청 시 Header에 'Authorization: Bearer <Access Token>' 형식으로 포함합니다."
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: "HttpOnly, SameSite=Strict refresh token cookie입니다. 운영 환경에서는 Secure 속성이 추가됩니다."
    wsTicket:
      type: apiKey
      in: query
      name: ticket
      description: "웹소켓 핸드셰이크에 사용하는 30초 TTL 일회용 티켓입니다."
  schemas:
...
    RefreshTokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: "새로 발급된 JWT 인증 토큰"
...
    ProblemDetails:
      type: object
      description: 에러 발생 시 반환되는 표준 형식 (RFC 9457)
      properties:
        type:
          type: string
          example: "about:blank"
        title:
          type: string
          example: "Bad Request"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "필수 값이 누락되었습니다."
  responses:
...
    Unauthorized:
      description: "인증 실패 (토큰 없음, 만료, 변조)"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    TooManyRequests:
      description: "요청 빈도 초과 (Rate Limit)"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...